Flask-Service/requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-CouchDB==0.2.1
Flask-Cors==3.0.10